Refactor Modifiers to bring them in line with actual ordering requirements.
Absolute ordering is not a requirement.  Relative ordering can be distilled to three rules that should be enforced throughout the lifetime of the Modifiers object.  Modifiers added with the "applyFirst" rule should always be invoked before those that are added with the "apply" rule.  Modifiers added with the "applyLast" rule should always be invoked after those added with the "apply" rule. Within each of these three subgroupings, modifiers are invoked based on FIFO ordering rules.